+++ HISTORY.rst
@@ -30,9 +30,9 @@ latest (unreleased)
   * Add a warehouse
   * Add locations WH and Stock
 * Configuration settings on main company
-
 * Install module 'Sensfly RMA MRP Repair'
 * Add warranty end date on stock production lot for serial numbers
+* Add Analytic Tags data: dimension and tags
 
 **Bugfixes**